+++ react/components/Border.tsx
@@ -1,14 +1,21 @@
 import React, { FunctionComponent } from 'react'
+import { useCssHandles } from 'vtex.css-handles'
+
 import StickyRow from './StickyRow'
 
 interface Props {
   sticky: boolean
 }
 
-const Border: FunctionComponent<Props> = ({ sticky }) => (
-  <StickyRow sticky={sticky}>
-    <div className="bb b--muted-3" />
-  </StickyRow>
-)
+const CSS_HANDLES = ['headerBorder'] as const
+
+const Border: FunctionComponent<Props> = ({ sticky }) => {
+  const handles = useCssHandles(CSS_HANDLES)
+  return (
+    <StickyRow sticky={sticky}>
+      <div className={`${handles.headerBorder} bb b--muted-3`} />
+    </StickyRow>
+  )
+}
 
 export default Border