+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- `info-card` component to allowed list `header-row`.
+
+
 ## [2.22.0] - 2019-10-16
 ### Added
 - New `forceCenterInnerContainer`, `headerBorder`, `headerRowContentContainer`, `headerSpacer`, `headerStickyRow` CSS handles.